Move keepBestChunks to batching.ts
